feat: build plugin with tsup
