+++ .github/workflows/main.yml
@@ -24,8 +24,5 @@ jobs:
       - name: Lint
         run: npm run lint
 
-      - name: Test
-        run: npm run test -- --ci --coverage --maxWorkers=2 --passWithNoTests
-
       - name: Build
         run: npm run build
