+++ tsconfig.json
@@ -29,7 +29,6 @@
     "skipLibCheck": true,
     // error out if import and file system have a casing mismatch. Recommended by TS
     "forceConsistentCasingInFileNames": true,
-    // `dts build` ignores this option, but it is commonly used when type-checking separately with `tsc`
-    "noEmit": true
+    "outDir": "dist"
   }
 }
